Changed print function to fi issue #63
Completely reworked how the print function works. before the only interactive printing (doing df) was an instruction to use df.frame to print it as a table or print(df) to print it as a dictionary. 

Considering we can just do df.df to and print(df.df) to print the dictionary, changed a couple of things. 
1) We made __str__ and __repr__ output the same thing so doing df interactively or print(df) will yield the same result. 
2) when using any of the print functions instead of printing to console line by line. it creates a string into a table variable and the _frame() methods returns that string. then doing __str__ or __repr__ calls on self._frame() and prints the output.

This allows us to test this as we can now check the return string of _frame against an expected output.

I also had to change the tests for __str__ and __repr__ to reflect this new functionality.